+++ script/api/api_city.cc
@@ -51,7 +51,7 @@ SQInteger world_get_city_by_index(HSQUIRRELVM vm)
 static void_t set_citygrowth(stadt_t *city, bool allow)
 {
 	static char param[16];
-	sprintf(param,"g%hi,%hi,%hi", city->get_pos().x, city->get_pos().y, allow );
+	sprintf(param,"g%hi,%hi,%hi", city->get_pos().x, city->get_pos().y, (short)allow );
 	werkzeug_t *wkz = werkzeug_t::simple_tool[WKZ_CHANGE_CITY_TOOL];
 	wkz->set_default_param( param );
 	wkz->flags |=  werkzeug_t::WFL_SCRIPT;
